[M152] [webgpu] Check Skia flush success before calling SetCleared
Original change's description:
> [webgpu] Check Skia flush success before calling SetCleared
>
> When using Skia to clear a shared image before, WebGPU decoder doesn't
> check if the Skia commands succeeded with Graphite. This CL fixes that
> by making the Graphite flush and submit functions return a success value
> and checking it in WebGPU decoder. Also add a unit test that simulates
> Graphite insert recording failure and assert that the shared image is
> not cleared incorrectly.
>
> Fixed: 517527943
> Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/8027742
> Commit-Queue: Bo Liu <boliu@chromium.org>
> Cr-Commit-Position: refs/heads/main@{#1677962}

(cherry picked from commit 79e5352b545898e6797e235996c43b0103d76dd9)

Bug: 546215696,517527943
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/8254886
Auto-Submit: chrome-cherry-picker@chops-service-accounts.iam.gserviceaccount.com <chrome-cherry-picker@chops-service-accounts.iam.gserviceaccount.com>
Commit-Queue: rubber-stamper@appspot.gserviceaccount.com <rubber-stamper@appspot.gserviceaccount.com>
Bot-Commit: rubber-stamper@appspot.gserviceaccount.com <rubber-stamper@appspot.gserviceaccount.com>
Cr-Commit-Position: refs/branch-heads/7977@{#1472}
Cr-Branched-From: b7fe14017379ddffae396d944fb8b59a5896c261-refs/heads/main@{#1669021}

+++ gpu/command_buffer/service/shared_image/wrapped_graphite_texture_holder.cc
@@ -21,7 +21,10 @@ WrappedGraphiteTextureHolder::WrappedGraphiteTextureHolder(
 WrappedGraphiteTextureHolder::~WrappedGraphiteTextureHolder() {
   auto destroy_resource = [](scoped_refptr<SharedContextState> context_state,
                              skgpu::graphite::BackendTexture texture) {
-    if (texture.isValid()) {
+    // It's possible in some tests that teardown with unsubmitted commands for
+    // the WrappedGraphiteTextureHolder to be released after the GraphiteDawn
+    // context has been nulled out and is in the middle of being torn down.
+    if (context_state->graphite_shared_context() && texture.isValid()) {
       context_state->graphite_shared_context()->deleteBackendTexture(texture);
     }
   };
